Refactored to be InvalidAccessTagCheck, moved from checks.validation.linear to checks.validation.tag, updated javadocs.
